Move SymbolicMatcher to tir/analysis
