+++ src/relax/transform/fuse_tir.cc
@@ -27,132 +27,11 @@
 #include <unordered_map>
 #include <unordered_set>
 
+#include "../../tir/analysis/symbolic_matcher.h"
 #include "../../tir/ir/functor_common.h"
 
 namespace tvm {
 namespace tir {
-
-// TODO(Siyuan): move it to somewhere under tir folder
-/*!
- * \brief Match symbolic vars according to the given PrimExpr, and update the var_remap.
- * Will throw errors if there is a mismatch.
- */
-class SymbolicMatcher : ExprFunctor<void(const PrimExpr& n, const PrimExpr& other)> {
- public:
-  explicit SymbolicMatcher(arith::Analyzer* analyzer, ffi::Map<tir::Var, PrimExpr>* var_remap)
-      : analyzer_(analyzer), var_remap_(var_remap) {}
-
-  void Match(const ffi::Array<PrimExpr>& params, const ffi::Array<PrimExpr>& args) {
-    CHECK_EQ(params.size(), args.size());
-    for (size_t i = 0; i < params.size(); ++i) {
-      Match(params[i], args[i]);
-    }
-  }
-  void Match(const PrimExpr& param, const PrimExpr& arg) {
-    VisitExpr(param, arg);
-    must_prove_ = analyzer_->Simplify(Substitute(must_prove_, *var_remap_));
-    CHECK(!is_zero(must_prove_));
-  }
-
- private:
-  void VisitExpr(const PrimExpr& node, const PrimExpr& other) {
-    if (node.same_as(other)) {
-      return;
-    } else if (node.dtype().code() != other.dtype().code()) {
-      LOG(FATAL) << "Parameter expression " << node << " with dtype " << node.dtype()
-                 << " cannot match to argument " << other << " with dtype " << other.dtype();
-    } else {
-      ExprFunctor::VisitExpr(node, other);
-    }
-  }
-
-#define TVM_DECLARE_SYMBOLIC_MATCHER_BINOP(OpName)                       \
-  void VisitExpr_(const OpName* op, const PrimExpr& other) {             \
-    const auto* rhs = other.as<OpName>();                                \
-    if (rhs) {                                                           \
-      VisitExpr(op->a, rhs->a);                                          \
-      VisitExpr(op->b, rhs->b);                                          \
-    } else {                                                             \
-      must_prove_ = must_prove_ && (ffi::GetRef<PrimExpr>(op) == other); \
-    }                                                                    \
-  }
-
-  TVM_DECLARE_SYMBOLIC_MATCHER_BINOP(AddNode);
-  TVM_DECLARE_SYMBOLIC_MATCHER_BINOP(SubNode);
-  TVM_DECLARE_SYMBOLIC_MATCHER_BINOP(MulNode);
-  TVM_DECLARE_SYMBOLIC_MATCHER_BINOP(DivNode);
-  TVM_DECLARE_SYMBOLIC_MATCHER_BINOP(ModNode);
-  TVM_DECLARE_SYMBOLIC_MATCHER_BINOP(EQNode);
-  TVM_DECLARE_SYMBOLIC_MATCHER_BINOP(NENode);
-  TVM_DECLARE_SYMBOLIC_MATCHER_BINOP(LTNode);
-  TVM_DECLARE_SYMBOLIC_MATCHER_BINOP(LENode);
-  TVM_DECLARE_SYMBOLIC_MATCHER_BINOP(GTNode);
-  TVM_DECLARE_SYMBOLIC_MATCHER_BINOP(GENode);
-  TVM_DECLARE_SYMBOLIC_MATCHER_BINOP(AndNode);
-  TVM_DECLARE_SYMBOLIC_MATCHER_BINOP(OrNode);
-  TVM_DECLARE_SYMBOLIC_MATCHER_BINOP(MinNode);
-  TVM_DECLARE_SYMBOLIC_MATCHER_BINOP(MaxNode);
-  TVM_DECLARE_SYMBOLIC_MATCHER_BINOP(FloorDivNode);
-  TVM_DECLARE_SYMBOLIC_MATCHER_BINOP(FloorModNode);
-
-  void VisitExpr_(const IntImmNode* op, const PrimExpr& other) {
-    const auto* rhs = other.as<IntImmNode>();
-    if (!rhs || (op->value != rhs->value)) {
-      LOG(FATAL) << "Parameter expression " << ffi::GetRef<PrimExpr>(op)
-                 << " expected an integer argument with value " << op->value << ", "
-                 << "but was provided with the argument " << other;
-    }
-  }
-
-  void VisitExpr_(const FloatImmNode* op, const PrimExpr& other) {
-    const auto* rhs = other.as<FloatImmNode>();
-    if (!rhs || (op->value != rhs->value)) {
-      LOG(FATAL) << "Parameter expression " << ffi::GetRef<PrimExpr>(op)
-                 << " expected an float argument with value " << op->value << ", "
-                 << "but was provided with the argument " << other;
-    }
-  }
-
-  void VisitExpr_(const CastNode* op, const PrimExpr& other) {
-    const auto* rhs = other.as<CastNode>();
-    if (!rhs) {
-      LOG(FATAL) << "Parameter expression " << ffi::GetRef<PrimExpr>(op) << " expected an cast to "
-                 << op->dtype << " as the argument, "
-                 << "but was provided with the argument " << other;
-    }
-    VisitExpr(op->value, rhs->value);
-  }
-
-  void VisitExpr_(const VarNode* op, const PrimExpr& rhs) {
-    auto lhs = ffi::GetRef<Var>(op);
-
-    if (lhs.same_as(rhs)) {
-      // Reference identity, no further checks needed.
-    } else if (op->dtype.code() != rhs->dtype.code()) {
-      LOG(FATAL) << "Parameter expression " << ffi::GetRef<PrimExpr>(op) << " with dtype "
-                 << op->dtype << " cannot match to argument " << rhs << " with dtype "
-                 << rhs.dtype();
-    } else if (auto it = var_remap_->find(lhs); it != var_remap_->end()) {
-      VisitExpr((*it).second, rhs);
-    } else {
-      var_remap_->Set(lhs, rhs);
-    }
-  }
-
-  void VisitExpr_(const SelectNode* op, const PrimExpr& other) {
-    const auto* rhs = other.as<SelectNode>();
-    if (rhs) {
-      VisitExpr(op->true_value, rhs->true_value);
-      VisitExpr(op->false_value, rhs->false_value);
-    } else {
-      must_prove_ = must_prove_ && (ffi::GetRef<PrimExpr>(op) == other);
-    }
-  }
-
-  arith::Analyzer* analyzer_;
-  ffi::Map<tir::Var, PrimExpr>* var_remap_;
-  PrimExpr must_prove_ = Bool(true);
-};
 
 /*!
  * \brief Substitute a given source buffer with a given target buffer in statements or expressions.
